Migrate to GitHub Pipelines from travis (#305)
* Update to include golangci-lint and go mod tidy

The go mod tidy to make sure the mod files are clean.

The golangci-lint helps in running in the linter locally instead just on
CI. The present linting has issues and that is the reason to include
only for new revisions.

Resolves #304

See also #303

* Implemented GitHub actions

Implemented GitHub actions for build, test, go mod tidy, coverage and
golangci-lint.

* Removed the .travis.yml

Removed the .travis.yml because of migration to GitHub actions.